Add FoodDatabaseItem model
