+++ AirFit/Modules/FoodTracking/Services/FoodVoiceServiceProtocol.swift
@@ -78,41 +78,3 @@ protocol FoodDatabaseServiceProtocol: Sendable {
     func searchFoods(query: String, limit: Int) async throws -> [FoodDatabaseItem]
 }
 
-// MARK: - Supporting Types
-struct FoodDatabaseItem: Identifiable, Sendable {
-    /// Unique identifier from the food database.
-    let id: String
-    /// Food name.
-    let name: String
-    /// Optional brand name.
-    let brand: String?
-    /// Calories contained in one serving.
-    let caloriesPerServing: Double
-    /// Protein in grams per serving.
-    let proteinPerServing: Double
-    /// Carbohydrates in grams per serving.
-    let carbsPerServing: Double
-    /// Fat in grams per serving.
-    let fatPerServing: Double
-    /// Nominal serving size amount.
-    let servingSize: Double
-    /// Unit for the serving size.
-    let servingUnit: String
-    /// Default quantity to prefill when adding.
-    let defaultQuantity: Double
-    /// Default unit to prefill when adding.
-    let defaultUnit: String
-
-    /// Human readable serving description.
-    var servingDescription: String {
-        "\(servingSize.formatted()) \(servingUnit)"
-    }
-}
-
-struct NutritionContext: Sendable {
-    let userPreferences: NutritionPreferences?
-    let recentMeals: [FoodItem]
-    let timeOfDay: Date
-}
-
-// MARK: - Food Voice Error
