+++ project.yml
@@ -167,6 +167,7 @@ targets:
       - AirFit/Modules/FoodTracking/Services/FoodVoiceServiceProtocol.swift
       - AirFit/Modules/FoodTracking/Services/NutritionService.swift
       - AirFit/Modules/FoodTracking/Services/FoodDatabaseService.swift
+      - AirFit/Modules/FoodTracking/Models/FoodDatabaseModels.swift
       - AirFit/Modules/FoodTracking/FoodTrackingCoordinator.swift
       - AirFit/Modules/FoodTracking/ViewModels/FoodTrackingViewModel.swift
       - AirFit/Modules/FoodTracking/Views/VoiceInputView.swift
